pyproject.toml
[project]
name = "rag-api"
version = "0.1.0"
description = "文档解析、向量化、混合检索与问答 API"
requires-python = ">=3.11,<4.0"
dependencies = [
    "fastapi>=0.104",
    "uvicorn[standard]>=0.24",
    "pydantic>=2.4",
    "pydantic-settings>=2.0",
    "python-dotenv>=1.0",
    "python-multipart>=0.0.9",
    "httpx>=0.27",
    "langchain>=0.2.14",
    "langchain-core>=0.2",
    "langchain-openai>=0.1.22",
    "langchain-text-splitters>=0.3",
    "llama-index-core>=0.12",
    "langgraph>=0.2.6",
    "langfuse>=3.0.0",
    "ragas>=0.2.0",
    "loguru>=0.7",
    "asyncpg>=0.29",
    "qdrant-client>=1.10",
    "openai>=1.0",
    "numpy>=1.26",
    "PyMuPDF>=1.23",
    "pdfplumber>=0.10",
    "python-docx>=1.1",
    "chardet>=5.2",
    "opencc-python-reimplemented>=0.1.7",
]

[project.optional-dependencies]
dev = ["mypy>=1.11", "ruff>=0.6", "pytest>=8.3"]

[build-system]
requires = ["setuptools>=73", "wheel"]
build-backend = "setuptools.build_meta"

[tool.ruff]
lint.select = ["E", "F", "I", "UP"]
lint.ignore = ["E501"]
